+++ prover/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod skyscraper_pow;
 pub mod skyscraper;
+pub mod skyscraper_traits_for_whir;
