+++ prover/src/main.rs
@@ -1,38 +1,30 @@
-use std::time::Instant;
-use std::borrow::Borrow;
-
-use ark_crypto_primitives::{
-    crh::{CRHScheme, TwoToOneCRHScheme},
-    merkle_tree::{
-        Config,
-        IdentityDigestConverter,
-    },
-    Error,
-};
-use nimue::plugins::ark::{FieldIOPattern, FieldReader, FieldWriter};
-
+use ark_crypto_primitives::merkle_tree::Config;
 use ark_ff::FftField;
 use ark_serialize::CanonicalSerialize;
-use nimue::{Arthur, DefaultHash, IOPattern, Merlin, ProofResult};
+use clap::Parser;
+use std::time::Instant;
+use nimue::{Arthur, IOPattern, Merlin};
+use prover::{
+    skyscraper::SkyscraperSponge, 
+    skyscraper_pow::SkyscraperPoW,
+    skyscraper_traits_for_whir::{
+        SkyscraperCRH, 
+        SkyscraperTwoToOne, 
+        SkyscraperMerkleConfig,
+    },
+};
 use whir::{
-    cmdline_utils::{AvailableFields, AvailableMerkle, WhirType},
     crypto::{
         fields::Field256,
         merkle_tree::{self, HashCounter},
     },
     parameters::*,
     poly_utils::{coeffs::CoefficientList, MultilinearPoint},
-    whir::Statement,
+    whir::{
+        fs_utils::{DigestReader, DigestWriter},
+        iopattern::DigestIOPattern,
+    },
 };
-
-use rand::Rng;
-
-use clap::Parser;
-use whir::whir::fs_utils::{DigestReader, DigestWriter};
-use whir::whir::iopattern::DigestIOPattern;
-use prover::skyscraper::{Skyscraper, SkyscraperSponge};
-use prover::skyscraper_pow::SkyscraperPoW;
-
 
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
@@ -78,93 +70,6 @@ fn main() {
     use merkle_tree::blake3 as mt;
 
     run_whir::<F, mt::MerkleTreeParams<F>>(args);
-}
-
-
-struct SkyscraperCRH;
-
-impl CRHScheme for SkyscraperCRH {
-    type Input = [Field256];
-    type Output = Field256;
-    type Parameters = ();
-
-    fn setup<R: Rng>(r: &mut R) -> Result<Self::Parameters, Error> {
-        Ok(())
-    }
-
-    fn evaluate<T: Borrow<Self::Input>>(
-        parameters: &Self::Parameters,
-        input: T,
-    ) -> Result<Self::Output, Error> {
-        let elems = input.borrow();
-        elems
-            .iter()
-            .cloned()
-            .reduce(prover::skyscraper::compress)
-            .ok_or(Error::IncorrectInputLength(0))
-    }
-}
-
-struct SkyscraperTwoToOne;
-
-impl TwoToOneCRHScheme for SkyscraperTwoToOne {
-    type Input = Field256;
-    type Output = Field256;
-    type Parameters = ();
-
-    fn setup<R: Rng>(r: &mut R) -> Result<Self::Parameters, Error> {
-        Ok(())
-    }
-
-    fn evaluate<T: Borrow<Self::Input>>(
-        _: &Self::Parameters,
-        left_input: T,
-        right_input: T,
-    ) -> Result<Self::Output, Error> {
-        Ok(prover::skyscraper::compress(
-            left_input.borrow().clone(),
-            right_input.borrow().clone(),
-        ))
-    }
-
-    fn compress<T: Borrow<Self::Output>>(
-        parameters: &Self::Parameters,
-        left_input: T,
-        right_input: T,
-    ) -> Result<Self::Output, Error> {
-        <Self as TwoToOneCRHScheme>::evaluate(parameters, left_input, right_input)
-    }
-}
-
-#[derive(Clone, Copy)]
-struct SkyscraperMerkleConfig;
-
-impl Config for SkyscraperMerkleConfig {
-    type Leaf = [Field256];
-    type LeafDigest = Field256;
-    type LeafInnerDigestConverter = IdentityDigestConverter<Field256>;
-    type InnerDigest = Field256;
-    type LeafHash = SkyscraperCRH;
-    type TwoToOneHash = SkyscraperTwoToOne;
-}
-
-impl DigestIOPattern<SkyscraperMerkleConfig> for IOPattern<SkyscraperSponge, Field256> {
-    fn add_digest(self, label: &str) -> Self {
-        <Self as FieldIOPattern<Field256>>::add_scalars(self, 1, label)
-    }
-}
-
-impl DigestWriter<SkyscraperMerkleConfig> for Merlin<SkyscraperSponge, Field256> {
-    fn add_digest(&mut self, digest: Field256) -> ProofResult<()> {
-        self.add_scalars(&[digest])
-    }
-}
-
-impl <'a> DigestReader<SkyscraperMerkleConfig> for Arthur<'a, SkyscraperSponge, Field256> {
-    fn read_digest(&mut self) -> ProofResult<Field256> {
-        let [r] = self.next_scalars()?;
-        Ok(r)
-    }
 }
 
 fn run_whir<F, MerkleConfig>(
